stepscreen code divided into seperate function files
